store/requirements.txt
Django==2.2.2
django-autofixture==0.12.1
django-extensions==2.1.7
django-extra-fields==1.2.2
django-filter==2.1.0
django-jalali==3.1.0
django-location-field==2.1.0
django-redis==4.10.0
django-redis-cache==2.0.0
djangorestframework==3.9.4
djangorestframework-gis==0.14
djangorestframework-simplejwt==4.3.0
jdatetime==3.2.0
Pillow==6.0.0
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.4
